Rename `location.py` -> `position.py`
`position.py` is a more expressive name
considering the content of the file
