+++ tests/units/test_mars_rover_forward.py
@@ -3,7 +3,7 @@
 import mars
 import pytest
 
-from location import CardinalPoint, Position
+from position import CardinalPoint, Position
 from rover import Rover
 from tests.helpers import assert_rover_state
